move ordering to separate fixture
