+++ examples/example05-runframe-for-cli.fixture.tsx
@@ -1,59 +1,13 @@
-import ky from "ky"
 import { RunFrameForCli } from "lib/components/RunFrameForCli/RunFrameForCli"
-import { useEventHandler } from "lib/components/RunFrameForCli/useEventHandler"
+import { RunFrameWithApi } from "lib/components/RunFrameWithApi/RunFrameWithApi"
 import { useRunFrameStore } from "lib/components/RunFrameWithApi/store"
-import { useEffect, useState } from "react"
+import { useState, useEffect } from "react"
 import { DebugEventsTable } from "./utils/DebugEventsTable"
+import { useEventHandler } from "lib/components/RunFrameForCli/useEventHandler"
 
 export default () => {
   const recentEvents = useRunFrameStore((state) => state.recentEvents)
   const pushEvent = useRunFrameStore((state) => state.pushEvent)
-  const [selectedScenario, setSelectedScenario] = useState<string>("none")
-  const [setSimulateScenarioOrder] = useRunFrameStore((s) => [
-    s.setSimulateScenarioOrder,
-  ])
-
-  const orderSteps = [
-    { key: "are_gerbers_generated", label: "Generate Gerbers" },
-    { key: "are_gerbers_uploaded", label: "Upload Gerbers" },
-    { key: "is_gerber_analyzed", label: "Analyze Gerber" },
-    { key: "are_initial_costs_calculated", label: "Calculate Initial Costs" },
-    { key: "is_pcb_added_to_cart", label: "Add PCB to Cart" },
-    { key: "is_bom_uploaded", label: "Upload BOM" },
-    { key: "is_pnp_uploaded", label: "Upload PnP" },
-    { key: "is_bom_pnp_analyzed", label: "Analyze BOM & PnP" },
-    { key: "is_bom_parsing_complete", label: "BOM Parsing Complete" },
-    { key: "are_components_available", label: "Components Available" },
-    { key: "is_patch_map_generated", label: "Generate Patch Map" },
-    { key: "is_json_merge_file_created", label: "Create JSON Merge File" },
-    { key: "is_dfm_result_generated", label: "Generate DFM Result" },
-    { key: "are_files_downloaded", label: "Download Files" },
-    {
-      key: "are_product_categories_fetched",
-      label: "Fetch Product Categories",
-    },
-    { key: "are_final_costs_calculated", label: "Calculate Final Costs" },
-    { key: "is_json_merge_file_updated", label: "Update JSON Merge File" },
-    { key: "is_added_to_cart", label: "Add to Cart" },
-  ]
-
-  useEffect(() => {
-    const pollInterval = setInterval(async () => {
-      try {
-        await ky
-          .get("/registry/_fake/move_orders_forward", {
-            headers: {
-              Authorization: `Bearer account-1234`,
-            },
-          })
-          .json()
-      } catch (error) {
-        console.error("Error polling order progress:", error)
-      }
-    }, 2000) // Poll every 2 seconds
-
-    return () => clearInterval(pollInterval)
-  }, [])
 
   useEventHandler(async (event) => {
     if (event.event_type === "REQUEST_TO_SAVE_SNIPPET") {
@@ -140,37 +94,7 @@ circuit.add(
 
   return (
     <>
-      <RunFrameForCli
-        debug
-        scenarioSelectorContent={
-          <div className="rf-flex rf-items-center">
-            <label
-              htmlFor="scenario"
-              className="rf-mr-2 rf-text-red-700 rf-text-sm"
-            >
-              Select Test Scenario:
-            </label>
-            <select
-              id="scenario"
-              className="rf-bg-white rf-border rf-border-gray-300 rf-rounded-md rf-shadow-sm rf-px-3 rf-py-1.5 rf-text-sm focus:rf-outline-none focus:rf-ring-2 focus:rf-ring-blue-500 focus:rf-border-blue-500"
-              value={selectedScenario}
-              onChange={(e) => {
-                setSelectedScenario(e.target.value)
-                setSimulateScenarioOrder(
-                  e.target.value === "none" ? "" : e.target.value,
-                )
-              }}
-            >
-              <option value="none">None</option>
-              {orderSteps.map((stage) => (
-                <option key={stage.key} value={stage.key}>
-                  {stage.label} Failed
-                </option>
-              ))}
-            </select>
-          </div>
-        }
-      />
+      <RunFrameForCli debug />
       <DebugEventsTable />
     </>
   )
